second build of the project
