+++ .github/workflows/main.yml
@@ -18,7 +18,6 @@ jobs:
 
     steps:
     - uses: actions/checkout@v3
-
     - name: Set up Python ${{ matrix.python-version }}
       uses: actions/setup-python@v4
       with:
@@ -77,7 +76,7 @@ jobs:
 #  deploy:
 #      runs-on: ubuntu-latest
 #      needs: build_and_push_to_docker_hub
-#      if: github.ref == 'refs/heads/master' && github.event_name == 'push'
+#      if: github.ref == 'refs/heads/main' && github.event_name == 'push'
 #
 #      steps:
 #        - name: executing remote ssh commands to deploy
@@ -110,7 +109,7 @@ jobs:
 #  send_message:
 #    runs-on: ubuntu-latest
 #    needs: deploy
-#    if: github.ref == 'refs/heads/master' && github.event_name == 'push'
+#    if: github.ref == 'refs/heads/main' && github.event_name == 'push'
 #    steps:
 #      - name: send message
 #        uses: appleboy/telegram-action@master
