+++ backend/config/urls.py
@@ -4,5 +4,5 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('api/', include('src.routers')),
+    path('api/', include('src.routers'))
 ]
