+++ backend/src/base/permissions.py
@@ -12,7 +12,7 @@ class IsAdminOrReadOnly(permissions.BasePermission):
     def has_permission(self, request, view):
         return (
             request.method in permissions.SAFE_METHODS
-            or (request.user.is_authenticated and request.user.is_superuser)
+            or (request.user.is_authenticated and request.user.is_staff)
         )
 
     def has_object_permission(self, request, view, obj):
