+++ backend/src/ingredients/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class IngredientsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'src.ingredients'
+    verbose_name = 'Ингредиенты'
