+++ backend/src/tags/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import Tag
+from src.tags.models import Tag
 
 
 class TagSerializer(serializers.ModelSerializer):
