+++ backend/src/tags/views.py
@@ -1,16 +1,15 @@
 from rest_framework import viewsets
 
-from .models import Tag
-from .serializers import TagSerializer
-from ..base.filters import RecipeFilter
-from ..base.permissions import IsAdminOrReadOnly
+from src.base.filters import RecipeFilter
+from src.base.permissions import IsAdminOrReadOnly
+from src.tags.models import Tag
+from src.tags.serializers import TagSerializer
 
 
 class TagViewSet(viewsets.ReadOnlyModelViewSet):
     """CRUD Тегов."""
-
     queryset = Tag.objects.all()
     serializer_class = TagSerializer
     permission_classes = (IsAdminOrReadOnly,)
-    pagination_class = None
     filter_class = RecipeFilter
+    pagination_class = None
