+++ infra/nginx.conf
@@ -9,28 +9,39 @@ server {
     listen 80;
     server_name 127.0.0.1;
     client_max_body_size 20m;
+
+    location /static/ {
+        root /var/html/;
+    }
+
     location /static/admin/ {
         root /var/html/;
     }
-    location /media/ {
-        root /var/html/;
-    }
+
     location /static/rest_framework {
         root /var/html/static/rest_framework;
     }
+
+    location /media/ {
+        root /var/html/;
+    }
+
+    location /admin/ {
+        proxy_pass http://backend:8000/admin/;
+    }
+
     location /api/ {
         proxy_set_header Host $host;
         proxy_set_header X-Forwarded-Host $host;
         proxy_set_header X-Forwarded-Server $host;
         proxy_pass http://backend:8000;
     }
-    location /admin/ {
-        proxy_pass http://backend:8000/admin/;
-    }
+
     location /api/docs/ {
         root /usr/share/nginx/html;
         try_files $uri $uri/redoc.html;
     }
+
     location / {
         root /usr/share/nginx/html;
         index index.html index.htm;
